Module Ocastore

Build your own distributed database! OCastore provides a database system which smoothly integrate with OCaml values to provide persistent storage and network distribution of your application state.

Using Ocastore, an application can directly manipulate standard OCaml types and values, with no impedance mismatch with the database model. The values stored in the database are exactly the OCaml values expected by your application, manipulated with regular OCaml code. OCaml types define the database schema, OCaml values are the database and OCaml is the query language.

This is loosely similar to using Marshal to store OCaml values on disk, but with granular control over the updates (instead of having to load and write the whole database state from / to disk) and with the ability to query the database over the network.

There are two libraries:

ocastore provides the basic building blocks to define your database schema, queries and database updates. This is the main module to use in your application code.

ocastore-eio to create or use a database stored on disk, or to serve and connect to a remote database over the network.

Database

type 'a schema = 'a Typerep_lib.Std.Typerep.t

The type of a schema describing the representation of an 'a value. This is automatically generated by using [@@deriving typerep] on your type definitions. See Schema PPX.

type 'a db

The type of a database storing a root value of type 'a.

val create_in_memory : 'a schema -> 'a -> 'a db

create_in_memory schema v returns a fresh in-memory database initialized to the state v described by the schema.

Example:

open Ocastore

type mydb = { what: string; ever: int } [@@deriving show, typerep]

let () =
  (* initialize eio and mirage-crypto *)
  Eio_main.run @@ fun env ->
  Mirage_crypto_rng_eio.run (module Mirage_crypto_rng.Fortuna) env @@ fun () ->

  let initial_value = { what = "fresh"; ever = 42 } in

  let db = create_in_memory typerep_of_mydb initial_value in

  Format.printf "%a@." pp_mydb (db_get db);
  assert (db_get db = initial_value);

  let new_value = { initial_value with what = "updated" } in
  db_set db new_value;
  Format.printf "%a@." (db_get db);
  assert (db_get db = new_value);

  db_close db

val db_close : 'a db -> unit

db_close db disconnects the database db. Further operations on the values stored by the database may fail.

val db_get : 'a db -> 'a

db_get db returns the current 'a value stored by the database db. Same as get db (db_root db).

val db_set : 'a db -> 'a -> unit

db_set db x updates the current value stored by the database db to be x. Same as set db (db_root db) x.

Links

To efficiently store and access very large values, it is necessary to introduce an indirection into the OCaml values manipulated by the database. This indirection permits lazy, on-demand, access to the database sub-values.

An 'a link is a pointer to an 'a value, which may reside in memory or on disk or can be acquired by a network call. They come with two operations: one to create a new 'a link to an 'a value, and one to fetch the 'a value pointed by an 'a link. Links behave like the identity type over 'a such that you may consider the equality 'a link ~= 'a to hold when reasonning about your program behaviour.

Links adds a serialization boundary to OCaml values: Their pointed value will only be read from disk/network on demand when calling fetch. By adding link annotations into OCaml types and values, you can control the granularity of the database access patterns.

For example:

type user = { name: string; avatar: Png.t link }
  [@@deriving typerep]

By adding a link to the avatar field of a user, we can avoid decoding it in queries that don't require it. We also avoid having to re-write the avatar on disk, when some other fields are updated.

Links are immutable and point to immutable values. Once created, they always point to the same value returned by fetch. As such, links also provide deduplication of the database values: If two different links are created to point to the same value, this value will only be stored once on disk. When the database is distributed over the network, clients exploit this to avoid fetching links that they have already obtained before.

Usage of mutable record fields, ref, array, Atomic.t etc is forbidden since those updates will not be seen by the OCastore database and will not be reflected on disk/network. See Mutable references to introduce granular mutations into your database.

type 'a link

The type of links pointing to an 'a value. The pointed 'a value can be obtained with fetch.

val link : 'a -> 'a link

link x returns a new link to the value x.

The address of a link to a value x is the content hash of the value x. As a result, two different links pointing to the same value will have the same address and will eventually be deduplicated. In practice, it means that the value x will be stored only once on disk and will be fetched only once over the network.

link (fetch l) ~= l

val fetch : 'a link -> 'a

fetch lnk returns the value pointed by the link lnk.

fetch (link x) ~= x

The loose equality ~= is only there to indicate that OCaml (=) builtin equality may not hold, even though the values are conceptually equal (and often even physically equal).

When designing your database model, you can start with standard OCaml types and values. Once you know your access patterns, you can introduce link annotations into your types to indicate that a sub-value is likely to survive database updates and/or will only rarely be accessed. This granularity will be exploited by Ocastore to update the disk and network clients more efficiently.

Since links are immutable, they encourage the use of purely functional datastructures. For example, you can adapt an existing datastructure by annotating each level of the recursion:

type t = s link
and s = Leaf | Bin of t * t
[@@deriving typerep]

Fixing the algorithms is then a matter of following the typechecker to introduce link and fetch appropriately. When updating such a datastructure, only the newly created links will need to be stored. This enables purely functional updates, which share a large part of the previous version, to run with the expected complexity.

TODO: Add stdlib-compatible implementations of Set, Map, etc.

Mutable references

To scale on multiple cores and on distributed systems, it is necessary to introduce mutable granularity instead of always updating the (highly contentious) root of the database. Since OCaml standard ref and other imperative mutations are unobservable by OCastore (which wouldn't be able to replicate the changes on disk or over the network), a new type is required to explicit mutations.

type 'a mut

The type of mutable references storing a value of type 'a.

val mut : 'a -> 'a mut

mut x returns a fresh mutable reference initialized to the value x.

val db_root : 'a db -> 'a mut

db_root db returns the mutable root of the database.

val get : _ db -> 'a mut -> 'a

get db m returns the current value of the reference m in the database db. Same as transaction db (read m).

The mutable reference m may exists in multiple databases, but with a different current value in each.

val set : _ db -> 'a mut -> 'a -> unit

set db m x updates the reference m to now hold the value x in the database db. Same as transaction db (write m x).

Transactions

To update or read multiple references all at once, in a linearizable way. This is the recommended way to query and update mutable references, since it offers more control over last-writer-win dataraces.

type tx

The type of a mutable transaction, reading and writing multiple mutable references.

val transaction : _ db -> (tx:tx -> 'a) -> 'a

transaction db fn runs the function fn ~tx in a new transaction tx over the database db. The updates of the transaction are made visible all at once. The function fn is expected to perform no other side-effects and to be quick: It may be called multiple times until it succeeeds (or aborts the transaction with an exception).

val read : tx:tx -> 'a mut -> 'a

read ~tx m returns the current value of the reference m inside the transaction tx.

val write : tx:tx -> 'a mut -> 'a -> unit

write ~tx m x updates the current value of the reference m inside the transaction tx. The change will only be visible by others once the transaction completes.

Reactive programming

Updates to mutable references can be observed by watchers, to provide reactive programming (e.g. to update the corresponding field in a GUI).

type watcher

The type of a watcher observing the changes made to a mutable reference.

val watch : _ db -> 'a mut -> ('a -> unit) -> watcher

watch db m fn will call fn for every new value affected to the mutable reference m in the database db. Returns a watcher that can be disabled with watch_stop when no longer needed.

val watch_stop : watcher -> unit

watch_stop w remove the watcher w. Its corresponding function will not be called anymore.

Schema PPX

Schemas describe the OCaml type definitions to the OCastore database. The schemas should be automatically generated by adding [@@deriving typerep] on the OCaml type definitions:

type t = ... [@@deriving typerep]

The deriving PPX automatically generates a value typerep_of_t : t schema (where the name t is the name you chose for your type), which can be given to the functions that create a database.

The following definitions are intended to be used by the PPX, not by users:

val typerep_of_link : 'a schema -> 'a link schema

typerep_of_link schema is the schema description of links pointing to an arbitrary 'a value described by schema.

val typerep_of_mut : 'a schema -> 'a mut schema

typerep_of_mut schema is the schema description of mutable references to an arbitrary 'a value described by schema.

You may have to open Typerep_lib.Std if the following aliases aren't sufficient:

val typerep_of_unit : unit schema

val typerep_of_bool : bool schema

val typerep_of_char : char schema

val typerep_of_int : int schema

val typerep_of_float : float schema

val typerep_of_string : string schema

val typerep_of_list : 'a schema -> 'a list schema

val typerep_of_array : 'a schema -> 'a array schema

val typerep_of_tuple2 : 'a schema -> 'b schema -> ('a * 'b) schema

val typerep_of_tuple3 : 'a schema -> 'b schema -> 'c schema -> ('a * 'b * 'c) schema

val typerep_of_tuple4 : 'a schema -> 'b schema -> 'c schema -> 'd schema -> ('a * 'b * 'c * 'd) schema

val typerep_of_tuple5 : 'a schema -> 'b schema -> 'c schema -> 'd schema -> 'e schema -> ('a * 'b * 'c * 'd * 'e) schema

val value_tuple0 : Typerep_lib.Std.tuple0

val typename_of_tuple0 : Typerep_lib.Std.tuple0 Typerep_lib.Std.Typename.t

val typerep_of_tuple0 : Typerep_lib.Std.tuple0 schema